+++ packages/expo-updates/CHANGELOG.md
@@ -130,6 +130,7 @@ _This version does not introduce any user-facing changes._
 
 ### 🐛 Bug fixes
 
+- [android] Fix update compatibility check to use `expoGoSDKVersion` instead of `runtimeVersion` in Expo Go update list. ([#44429](https://github.com/expo/expo/pull/44429) by [@fschindler](https://github.com/fschindler))
 - Pass absolute path to CLI helpers when creating build manifest, since the underlying functions now handle entry file inputs properly, instead of applying `mainModuleName` semantics to them ([#44414](https://github.com/expo/expo/pull/44414) by [@kitten](https://github.com/kitten))
 - [ios] Fix loading assets in brownfield ([#44724](https://github.com/expo/expo/pull/44724) by [@gabrieldonadel](https://github.com/gabrieldonadel))